adding database support
we basically need two tables. the first one to store all the files that has been
found together with their sizes and a second one to store their hashes. maybe
this can be made more efficient later. right now i am not in the mood to think
much about efficiency of the database layout.